+++ tests/test_utils.py
@@ -14,7 +14,27 @@ class OptionsTestCase(unittest.TestCase):
             ),
             with_underscore=dict(
                 options=OrderedDict(foo_baz='bar'),
-                expected_str_version='--foo-baz bar',
+                expected_str_version='--foo_baz bar',
+            ),
+            multiword=dict(
+                options=OrderedDict(foo='bar baz'),
+                expected_str_version='--foo "bar baz"',
+            ),
+            empty=dict(
+                options=OrderedDict(foo=''),
+                expected_str_version='--foo ""',
+            ),
+            with_single_quotes=dict(
+                options=OrderedDict(foo="'bar'"),
+                expected_str_version='--foo "\'bar\'"',
+            ),
+            with_double_quotes=dict(
+                options=OrderedDict(foo='"bar"'),
+                expected_str_version='--foo "\\"bar\\""',
+            ),
+            with_quotes_and_spaces=dict(
+                options=OrderedDict(foo='"bar" \'baz\''),
+                expected_str_version='--foo "\\"bar\\" \'baz\'"',
             ),
             single_length=dict(
                 options=OrderedDict(foo='bar'),